requirements.txt
torch
torchvision
matplotlib
numpy

flask
Pillow
flask_cors